tools/install/pyproject.toml
[project]
name = "rp1-opencode"
version = "4.0.0"
description = "Installation tool for managing rp1 plugins on OpenCode platform"
authors = [{ name = "Prem Pillai", email = "prem.pillai@gmail.com" }]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.13"
keywords = ["opencode", "rp1", "installation", "plugin-manager"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Build Tools",
  "Topic :: System :: Installation/Setup",
]

dependencies = [
  "typer>=0.12.0",
  "rich>=13.7.0",
  "returns>=0.23.0",
  "pyyaml>=6.0",
]

[[tool.uv.index]]
name = "pypi"
url = "https://pypi.org/simple"
default = true

[tool.uv.pip]
index-url = "https://pypi.org/simple"
extra-index-url = []

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.1.0",
  "pytest-mock>=3.12.0",
  "mypy>=1.8.0",
  "ruff>=0.6.0",
  "types-pyyaml>=6.0",
]

[project.scripts]
rp1-opencode = "rp1_opencode.cli:app"

[project.urls]
Homepage = "https://github.com/rp1-run/rp1"
Repository = "https://github.com/rp1-run/rp1"
Issues = "https://github.com/rp1-run/rp1/issues"
Documentation = "https://github.com/rp1-run/rp1/tree/main/docs/opencode"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rp1_opencode"]
include = ["src/rp1_opencode/assets/*.txt"]

[tool.hatch.build.targets.wheel.force-include]
"dist/opencode" = "rp1_opencode/artifacts"

[tool.ruff]
line-length = 120
target-version = "py313"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4", # flake8-comprehensions
  "UP", # pyupgrade",
]
ignore = [
  "E501", # line too long (handled by formatter)
  "B008", # function call in argument defaults (standard Typer pattern)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
strict = true
plugins = ["returns.contrib.mypy.returns_plugin"]
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "--strict-markers",
  "--strict-config",
  "--cov=src/rp1_opencode",
  "--cov-report=term-missing",
  "--cov-report=html",
  "--cov-branch",
]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "class .*\\bProtocol\\):",
  "@(abc\\.)?abstractmethod",
]
